+++ Pulsarc/UI/Screens/Gameplay/UI/JudgeBox.cs
@@ -6,7 +6,7 @@ namespace Pulsarc.UI.Screens.Gameplay.UI
 {
     class JudgeBox : Drawable
     {
-        List<KeyValuePair<long, int>> toDraw;
+        List<KeyValuePair<double, int>> toDraw;
         Dictionary<int, Judge> judges;
 
         public JudgeBox(Vector2 position) : base(Skin.defaultTexture)
@@ -21,15 +21,15 @@ namespace Pulsarc.UI.Screens.Gameplay.UI
                 judges.Add(judge.score,new Judge(judge.score, new Vector2(position.X, position.Y)));
             }
 
-            toDraw = new List<KeyValuePair<long, int>>();
+            toDraw = new List<KeyValuePair<double, int>>();
         }
 
-        public void Add(long time, int judgeKey)
+        public void Add(double time, int judgeKey)
         {
-            toDraw.Add(new KeyValuePair<long, int>(time, judgeKey));
+            toDraw.Add(new KeyValuePair<double, int>(time, judgeKey));
         }
 
-        public void Update(long time)
+        public void Update(double time)
         {
             int judgeDisplayTimeMs = 100;
 
@@ -45,7 +45,7 @@ namespace Pulsarc.UI.Screens.Gameplay.UI
 
         public override void Draw()
         {
-            foreach(KeyValuePair<long,int> judge in toDraw)
+            foreach(KeyValuePair<double, int> judge in toDraw)
             {
                 judges[judge.Value].Draw();
             }
